+++ src/display/engine/rules/Shape.java
@@ -5,7 +5,7 @@ public abstract class Shape extends GraphicalObject {
     public Shape(
         int posX, int posY,
         int scaleX, int scaleY,
-        int red, int blue, int green
+        float red, float blue, float green
     ) {
         super(posX, posY, scaleX, scaleY, red, blue, green);
     }
