+++ src/display/engine/shapes/Rectangle.java
@@ -6,7 +6,7 @@ public class Rectangle extends Shape {
     public Rectangle(
         int posX, int posY,
         int scaleX, int scaleY,
-        int red, int blue, int green
+        float red, float blue, float green
     ) {
         super(posX, posY, scaleX, scaleY, red, blue, green);
     }
